Quick hack to show PBM images
requires the netpbm library (http://netpbm.sourceforge.net)

This patch introduces the -i parameter.
It allows specifying a file to read PBM images from.
It is possible to concatenate up to 10 images in that file.
The image index in that file can be appended to a percentage with '@'.
```@0``` will disable the image again.
Black pixels in the image will invert bar and background color.

Example:
```
    cat extra/volume.pbm extra/brightness.pbm >icons.pbm
    ./build/wob -i icons.pbm
    10 @1
    20
    15 @0
    5
    10 @2
    25
```